Add xtrace-mig support to mig
MIG now supports a third (or fifth, dependning on how you count) output type,
xtrace-mig. This is a C file that is able to parse serialized binary messages of
this subsystem and pass information about their arguments and return values to
the given callbacks.

This *XtraceMig.c file is not linked into either clients or servers; instead it
is compiled as a dynamic library and placed into /usr/lib/darling/xtrace-mig/

This will be used by the xtrace tool to be able to parse and display MIG routine
calls.