Remove platform-include and kernel-include and unify all headers into the SDK in Developer [skip ci]
The SDK include directory has been recreated to match the structure of macOS's actual SDK include directory.

The SDK include directory symlinks as many headers as possible, and provides copies of the open source (and therefore freely redistributable) headers found in macOS's SDK that can't be found elsewhere.

The CMakeLists will be updated in a later commit to include this directory instead of platform-include, and to resolve any missing header errors.