CoreServices: Add BackupCore and MDItem stubs
MDItem's current code is just a bunch of stubs. However, BackupCore's current code is technically implemented (Darling doesn't do backups, so it just tells users that everything is excluded from backups).

Note that I didn't verify the values of each and every MDItem constant, but those kinds of constants usually follow a pattern, and in this case, the pattern seemed to be that their string content is just a copy of their variable names. (I could probably have used a macro to do it, but *shrug*, my editor made it easy).