Use sys_kevent64 instead of kevent64 in kevent_qos
I'm not sure how/why the previous code was working (I guess we weren't running into error codes very often), but I found that it wouldn't work properly when errors (such as EINTR) occurred. The issue was that `kevent64` is the libc wrapped version of the call that returns its error codes in `errno`, so we were only getting `-1` for all errors, and then our own libc wrapper was translating that into EPERM in our errno return. What we actually want is the raw error value, so we have to use `sys_kevent64` directly.