Set TMPDIR with shellspawn
macOS always sets it too.

Discovered via Tokio's tests: Rust tries to use `TMPDIR` and falls back to `/tmp`. Tokio's tests compare that two paths are equal: the one used to create a Unix socket, and the one returned as the peer address of that socket.
Due to our vchroot translation code, the symlinks are resolved in the returned peer path and Tokio complains that the paths aren't equal.

We can workaround this in this case by setting TMPDIR to a non-symlinked path. However, there might be other cases in other software where our vchroot translation (from Linux paths to Darwin paths) follows symlinks and returns a different path name than was expected, so be wary of this a potential issue in the future.