Quite a few syscall updates; working execve; new tool
  * Many syscalls have been updated to use darlingserver RPC calls instead of LKM calls.
  * The elfcalls threading glue now performs checkins and checkouts for non-main threads.
  * mldr's main stack allocation logic has been fixed to not overwrite the commpage.
  * execve now works properly with mldr and informs the server of the success or failure of the execve call (using some close-on-exec FD trickery; see the corresponding commit in darlingserver for more details).

One very useful additional that merits its own paragraph is the new `darling-coredump` tool. binfmt_misc provides no way to register a coredump handler, and trying to intercept all the possible signals that can generate a coredump in-process and handle them appropriately is more trouble than it's worth. Therefore, I created a new tool that converts a kernel-generated ELF coredump into a proper Mach-O coredump. The tool is currently only compatible with kernel-generated coredumps; coredumps produced with `gcore` seem to be missing certain information that causes the tool to produce an invalid coredump (and I'm not quite sure what is wrong with it).