mldr: Unmap reservation before mapping in actual contents
This is necessary as a result of the changes from e9837a743513a001a95e4fc1c5451cd6a916255f.
We do want to make sure we don't replace any previous memory we didn't
load ourselves, so in order to keep that, we just unmap the memory we reserve,
since there's no race condition like there could have been when we were
doing this in the LKM. No one else can be allocating memory in our
process but us.