Add a sigaltstack guard by default and change default size to 16KiB
It turns out that in some cases (with nested signals) the default
sigaltstack of 8KiB is not enough; 16KiB seems to be enough
for these cases (for now).

I suspected there was some stack corruption/overflow going on, so I
added a guard page for the default handler and found that, yes, we were
overflowing the default sigstack. That's why I've left the guard page
enabled by default now for sigstacks; if we do ever overflow it, we'll
know right away (rather than with some seemingly unrelated error later
on).