Add .dylib files (symbolic-links) for generate-offlineasm.sh
The script relies on the System libraries existing in the SDK, otherwise, the linker will fail. For some reason, some of these files are ignored by the .gitignore file. I'm not sure why that is the case...

In an official macOS SDK, .tbd (a text-based library stub) files are used instead. However, I wasn't able to figure out if these files can be used on Linux. So for the time being, I'll just symbol-link the .dylib files when needed.