Update CMAKE_OSX_DEPLOYMENT_TARGET To 11.0
A bit of a backstory: macOS Big Sur was the first version of macOS to increment it's major version number from 10 to 11. Since Apple doesn't want to break old applications that assume the "10.x" version scheme, they applied a small hack in the xnu kernel to have older applications read off of the SystemVersionCompat.plist instead. So instead of seeing "11.6", the app would see "10.16" instead.

Since the macOS deployment target was not updated to BIg Sur's version, sw_vers would report "10.16". Fixes #1334