Set Fall Back Arch To The Primary Arch For MIG
I was looking through the build scripts in XNU to see how MIG is being used. While there are some cases where we need to explicitly set the arch to its 32-bit equivalent (ex: i386 instead of x86-64), those situations seem to be the exception rather than the norm.

As a result, I don't think we should be hardcoding i386 as a fallback. We should use the primary architecture instead.